refactor: auth-context 폴더 위치 수정
