+++ src/features/posts/ui/post-list-item.tsx
@@ -4,7 +4,7 @@ import CommentIcon from '@/assets/icons/comment.svg';
 import EyeIcon from '@/assets/icons/eye.svg';
 import HeartIcon from '@/assets/icons/mdi_heart.svg';
 import { type Post } from '@/entities/posts';
-import { useAuthContext } from '@/shared/context/auth-context';
+import { useAuthContext } from '@/features/auth/context/auth-context';
 import { isValidUrl } from '@/shared/lib/is-valid-url';
 import Image from 'next/image';
